requirements.txt
pandas~=2.2.3
numpy~=2.1.3
scikit-learn~=1.6.1
pymongo~=4.12.0
python-dotenv~=1.1.0
tensorflow==2.13.0
requests==2.31.0
fastapi~=0.115.12
uvicorn~=0.34.0
pydantic~=2.11.3